+++ src/utilsRefactored/seqFunctions.cpp
@@ -4,12 +4,13 @@
 namespace RefactorProcessing {
 
     // encode a k-mer into an integer hash, return -1 if the seq length is less than kMerSize, or -i-2 if the i-th character is invalid
-    int64_t encodeKmer(const std::string_view &seq, int kMerSize) {
+    int64_t encodeKmer(const std::string_view &seq, unsigned int kMerSize) {
+        // TODO: `kMerSize` should always be unsigned!
         if (seq.length() < kMerSize) {
             return -1;
         }
         int64_t hash = 0;
-        for (int i = 0; i < kMerSize; ++i) {
+        for (int i = 0; i < (int) kMerSize; ++i) {
             int32_t idx = charToIndex(seq[i]);
             if (idx < 0) return -i - 2;
             hash = (hash << 2) | idx;
